Major changes to Texture, FrameBuffer Objects, Bitmap Loading,  Camera Zoom Fixed

+++ CreatureEngine/Core/Application.cpp
@@ -514,7 +514,7 @@ LRESULT CALLBACK WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
 
 	case WM_SIZE:
 	{
-		Vec2 sz = SplitLParam(lParam);
+		Vec2 sz = SplitLParam((int)lParam);
 		Application::get().Resize({ sz.x,sz.y });
 	}break;
